Simplify the AbstractNetworkTest tree (directed branch): Delete StandardSimpleDirectedNetworkTest
This is similar to an earlier refector of the AbstractGraphTest tree.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=300062171
