+++ android/guava-tests/test/com/google/common/graph/StandardDirectedMultiNetworkTest.java
@@ -26,7 +26,8 @@ import org.junit.runners.JUnit4;
 
 /** Tests for a directed {@link StandardMutableNetwork} allowing parallel edges and self-loops. */
 @RunWith(JUnit4.class)
-public class StandardDirectedMultiNetworkTest extends StandardDirectedNetworkTest {
+public class StandardDirectedMultiNetworkTest extends StandardMutableDirectedNetworkTest {
+
   @Override
   MutableNetwork<Integer, String> createGraph() {
     return NetworkBuilder.directed().allowsParallelEdges(true).allowsSelfLoops(true).build();
