+++ android/guava-tests/test/com/google/common/graph/StandardSimpleDirectedSortedNetworkTest.java
@@ -30,7 +30,7 @@ import org.junit.runners.JUnit4;
  * against a sorted graph so as to cover {@link MapRetrievalCache}.
  */
 @RunWith(JUnit4.class)
-public class StandardSimpleDirectedSortedNetworkTest extends StandardSimpleDirectedNetworkTest {
+public class StandardSimpleDirectedSortedNetworkTest extends AbstractStandardDirectedNetworkTest {
 
   @Override
   MutableNetwork<Integer, String> createGraph() {
@@ -40,6 +40,16 @@ public class StandardSimpleDirectedSortedNetworkTest extends StandardSimpleDirec
         .edgeOrder(sorted(Ordering.natural()))
         .nodeOrder(sorted(Ordering.natural()))
         .build();
+  }
+
+  @Override
+  void addNode(Integer n) {
+    networkAsMutableNetwork.addNode(n);
+  }
+
+  @Override
+  void addEdge(Integer n1, Integer n2, String e) {
+    networkAsMutableNetwork.addEdge(n1, n2, e);
   }
 
   @Override
