+++ guava-tests/test/com/google/common/graph/AbstractStandardDirectedNetworkTest.java
@@ -20,6 +20,7 @@ import static com.google.common.graph.GraphConstants.ENDPOINTS_MISMATCH;
 import static com.google.common.graph.TestUtil.assertEdgeNotInGraphErrorMessage;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.TruthJUnit.assume;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import com.google.common.collect.ImmutableSet;
@@ -30,13 +31,9 @@ import org.junit.After;
 import org.junit.Test;
 
 /**
- * Abstract base class for testing implementations of {@link Network} interface.
- *
- * <p>This class is responsible for testing that a directed implementation of {@link Network} is
- * correctly handling directed edges. Implementation-dependent test cases are left to subclasses.
- * Test cases that do not require the graph to be directed are found in superclasses.
+ * Abstract base class for testing directed {@link Network} implementations defined in this package.
  */
-public abstract class AbstractDirectedNetworkTest extends AbstractNetworkTest {
+public abstract class AbstractStandardDirectedNetworkTest extends AbstractNetworkTest {
 
   @After
   public void validateSourceAndTarget() {
@@ -59,6 +56,144 @@ public abstract class AbstractDirectedNetworkTest extends AbstractNetworkTest {
         assertThat(node.equals(adjacentNode) || Collections.disjoint(edges, antiParallelEdges))
             .isTrue();
       }
+    }
+  }
+
+  @Override
+  @Test
+  public void nodes_checkReturnedSetMutability() {
+    Set<Integer> nodes = network.nodes();
+    try {
+      nodes.add(N2);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addNode(N1);
+      assertThat(network.nodes()).containsExactlyElementsIn(nodes);
+    }
+  }
+
+  @Override
+  @Test
+  public void edges_checkReturnedSetMutability() {
+    Set<String> edges = network.edges();
+    try {
+      edges.add(E12);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.edges()).containsExactlyElementsIn(edges);
+    }
+  }
+
+  @Override
+  @Test
+  public void incidentEdges_checkReturnedSetMutability() {
+    addNode(N1);
+    Set<String> incidentEdges = network.incidentEdges(N1);
+    try {
+      incidentEdges.add(E12);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.incidentEdges(N1)).containsExactlyElementsIn(incidentEdges);
+    }
+  }
+
+  @Override
+  @Test
+  public void adjacentNodes_checkReturnedSetMutability() {
+    addNode(N1);
+    Set<Integer> adjacentNodes = network.adjacentNodes(N1);
+    try {
+      adjacentNodes.add(N2);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.adjacentNodes(N1)).containsExactlyElementsIn(adjacentNodes);
+    }
+  }
+
+  @Override
+  public void adjacentEdges_checkReturnedSetMutability() {
+    addEdge(N1, N2, E12);
+    Set<String> adjacentEdges = network.adjacentEdges(E12);
+    try {
+      adjacentEdges.add(E23);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N2, N3, E23);
+      assertThat(network.adjacentEdges(E12)).containsExactlyElementsIn(adjacentEdges);
+    }
+  }
+
+  @Override
+  @Test
+  public void edgesConnecting_checkReturnedSetMutability() {
+    addNode(N1);
+    addNode(N2);
+    Set<String> edgesConnecting = network.edgesConnecting(N1, N2);
+    try {
+      edgesConnecting.add(E23);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.edgesConnecting(N1, N2)).containsExactlyElementsIn(edgesConnecting);
+    }
+  }
+
+  @Override
+  @Test
+  public void inEdges_checkReturnedSetMutability() {
+    addNode(N2);
+    Set<String> inEdges = network.inEdges(N2);
+    try {
+      inEdges.add(E12);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.inEdges(N2)).containsExactlyElementsIn(inEdges);
+    }
+  }
+
+  @Override
+  @Test
+  public void outEdges_checkReturnedSetMutability() {
+    addNode(N1);
+    Set<String> outEdges = network.outEdges(N1);
+    try {
+      outEdges.add(E12);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.outEdges(N1)).containsExactlyElementsIn(outEdges);
+    }
+  }
+
+  @Override
+  @Test
+  public void predecessors_checkReturnedSetMutability() {
+    addNode(N2);
+    Set<Integer> predecessors = network.predecessors(N2);
+    try {
+      predecessors.add(N1);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(network.predecessors(N2)).containsExactlyElementsIn(predecessors);
+    }
+  }
+
+  @Override
+  @Test
+  public void successors_checkReturnedSetMutability() {
+    addNode(N1);
+    Set<Integer> successors = network.successors(N1);
+    try {
+      successors.add(N2);
+      fail(ERROR_MODIFIABLE_COLLECTION);
+    } catch (UnsupportedOperationException e) {
+      addEdge(N1, N2, E12);
+      assertThat(successors).containsExactlyElementsIn(network.successors(N1));
     }
   }
 
@@ -273,5 +408,40 @@ public abstract class AbstractDirectedNetworkTest extends AbstractNetworkTest {
     } catch (IllegalArgumentException e) {
       assertThat(e).hasMessageThat().contains(ENDPOINTS_MISMATCH);
     }
+  }
+
+  @Test
+  public void addEdge_selfLoop() {
+    assume().that(graphIsMutable()).isTrue();
+
+    try {
+      networkAsMutableNetwork.addEdge(N1, N1, E11);
+      fail(ERROR_ADDED_SELF_LOOP);
+    } catch (IllegalArgumentException e) {
+      assertThat(e).hasMessageThat().contains(ERROR_SELF_LOOP);
+    }
+  }
+
+  /**
+   * This test checks an implementation dependent feature. It tests that the method {@code addEdge}
+   * will silently add the missing nodes to the graph, then add the edge connecting them. We are not
+   * using the proxy methods here as we want to test {@code addEdge} when the end-points are not
+   * elements of the graph.
+   */
+  @Test
+  public void addEdge_nodesNotInGraph() {
+    assume().that(graphIsMutable()).isTrue();
+
+    networkAsMutableNetwork.addNode(N1);
+    assertTrue(networkAsMutableNetwork.addEdge(N1, N5, E15));
+    assertTrue(networkAsMutableNetwork.addEdge(N4, N1, E41));
+    assertTrue(networkAsMutableNetwork.addEdge(N2, N3, E23));
+    assertThat(network.nodes()).containsExactly(N1, N5, N4, N2, N3).inOrder();
+    assertThat(network.edges()).containsExactly(E15, E41, E23).inOrder();
+    assertThat(network.edgesConnecting(N1, N5)).containsExactly(E15);
+    assertThat(network.edgesConnecting(N4, N1)).containsExactly(E41);
+    assertThat(network.edgesConnecting(N2, N3)).containsExactly(E23);
+    // Direction of the added edge is correctly handled
+    assertThat(network.edgesConnecting(N3, N2)).isEmpty();
   }
 }
